refactor: [FE] 로비페이지 소켓 코드 훅으로 분리
